Shooting
I finally got the player to shoot using the space bar and got the
bullets to spawn and shoot from the player, as well as being able to
fire multiple shots. However, the shots only move to the right for now
and I still need to dispose of them to prevent memory leaks.